perf: :memo: add metrics docs
