+++ docs/metrics/metrics.md
@@ -1,0 +1,1 @@
+## Getting Started
